Attempt at Implementation of State 4
As of now I am running into a "list out of range" error, I think I need to rethink how I approach this problem.

Right now if state 3 runs into red, it doesn't now what to do, and I lazily sent it to state 4 to see if it would miraculously work, it did not. I need to rethink